visitors can still view comments

+++ client/src/components/CommentItem.jsx
@@ -1,6 +1,7 @@
 import React, { useState } from "react";
 import Moment from "react-moment";
 import { connect } from "react-redux";
+import { Link } from "react-router-dom";
 import { deleteComment, addHeart, removeHeart } from "../actions/post-action";
 import PropTypes from "prop-types";
 
@@ -15,32 +16,53 @@ const CommentItem = ({
   const heartsUsers = comment.hearts.map((heart) => heart.user);
   console.log("heartsusers:", heartsUsers);
   console.log("Auth:", auth);
-  return (
-    <div className="comment-item">
-      <div className="comment-header">
-        <p style={{ textDecoration: "underline" }}>{comment.name}</p>
-        {!auth.loading && auth.user._id === comment.user && (
-          <button onClick={() => deleteComment(post._id, comment._id)}>
-            X
+
+  if (!auth.isAuthenticated) {
+    return (
+      <div className="comment-item">
+        <div className="comment-header">
+          <p style={{ textDecoration: "underline" }}>{comment.name}</p>
+        </div>
+        <Moment format="MM/DD/YYYY">{comment.date}</Moment>
+        <p>{comment.text}</p>
+        <Link to="/login">
+          <button>
+            <i className="far fa-heart">
+              {comment.hearts.length > 0 && comment.hearts.length}
+            </i>
+          </button>
+        </Link>
+      </div>
+    );
+  }
+  if (auth.isAuthenticated) {
+    return (
+      <div className="comment-item">
+        <div className="comment-header">
+          <p style={{ textDecoration: "underline" }}>{comment.name}</p>
+          {!auth.loading && auth.user._id === comment.user && (
+            <button onClick={() => deleteComment(post._id, comment._id)}>
+              X
+            </button>
+          )}
+        </div>
+        <Moment format="MM/DD/YYYY">{comment.date}</Moment>
+        <p>{comment.text}</p>
+        {!auth.loading && heartsUsers.includes(auth.user._id) && (
+          <button onClick={() => removeHeart(post._id, comment._id)}>
+            <i className="fas fa-heart">{comment.hearts.length}</i>
+          </button>
+        )}
+        {!auth.loading && !heartsUsers.includes(auth.user._id) && (
+          <button onClick={() => addHeart(post._id, comment._id)}>
+            <i className="far fa-heart">
+              {comment.hearts.length > 0 && comment.hearts.length}
+            </i>
           </button>
         )}
       </div>
-      <Moment format="MM/DD/YYYY">{comment.date}</Moment>
-      <p>{comment.text}</p>
-      {!auth.loading && heartsUsers.includes(auth.user._id) && (
-        <button onClick={() => removeHeart(post._id, comment._id)}>
-          <i className="fas fa-heart">{comment.hearts.length}</i>
-        </button>
-      )}
-      {!auth.loading && !heartsUsers.includes(auth.user._id) && (
-        <button onClick={() => addHeart(post._id, comment._id)}>
-          <i className="far fa-heart">
-            {comment.hearts.length > 0 && comment.hearts.length}
-          </i>
-        </button>
-      )}
-    </div>
-  );
+    );
+  }
 };
 
 CommentItem.propTypes = {
